requirements.txt
influxdb~=5.3.1
PyYAML~=5.3.1